Split main window and entry point into two files
